+++ src/rewiews/filters/filterlist.js
@@ -1,0 +1,10 @@
+'use strict';
+
+/** @enum {string} */
+module.exports = {
+  ALL: 'reviews-all',
+  BAD: 'reviews-bad',
+  GOOD: 'reviews-good',
+  POPULAR: 'reviews-popular',
+  RECENT: 'reviews-recent'
+};
